Consolidate LSA_OBJECT_ATTRIBUTES and LSA_UNICODE_STRING
