--- src/Common/src/Interop/Windows/SspiCli/Interop.UNICODE_STRING.cs
@@ -1,18 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Runtime.InteropServices;
-
-internal static partial class Interop
-{
-    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-    internal struct UNICODE_STRING
-    {
-        internal ushort Length;
-        internal ushort MaximumLength;
-        [MarshalAs(UnmanagedType.LPWStr)]
-        internal string Buffer;
-    }
-}
